+++ NEWS.md
@@ -28,6 +28,8 @@
 - Fixed file-type icons not displaying in Finder on Mac (#12252)
 - Fixed saving and restoring window location when maximized or partially offscreen (#12463)
 - Fixed display of macOS message dialogs (#12928)
+- Set theme of menu bar, title bar, and dialogs (dark vs. light) based on RStudio theme (#12247)
+- Fixed issues with mouse back / forward navigation in Source pane, Help pane (#12932)
 
 #### Posit Workbench
 - Fixed unlabeled buttons for screen reader users when page is narrow [Accessibility] (rstudio/rstudio-pro#4340)